Fix MSVC warning C4244: '=': conversion from 'const int' to 'uint8_t', possible loss of data

+++ src/template.cpp
@@ -408,7 +408,7 @@ DROID_TEMPLATE::DROID_TEMPLATE()  // This constructor replaces a memset in scrAs
 	, stored(false)
 	, enabled(false)
 {
-	std::fill_n(asParts, DROID_MAXCOMP, 0);
+	std::fill_n(asParts, DROID_MAXCOMP, static_cast<uint8_t>(0));
 	std::fill_n(asWeaps, MAX_WEAPONS, 0);
 }
